Rearrange unittest and travis config
