+++ deployment/compose.env.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+
+export COMPOSE_PROJECT_NAME=headless
