+++ src/scripts/start_coverage_unittest_worker.sh
@@ -1,0 +1,8 @@
+#!/usr/bin/env bash
+
+if [ -z "${TASK_ALWAYS_EAGER}" ] || [ "${TASK_ALWAYS_EAGER}" = "False" ]; then
+	for i in `seq 1 ${WORKER_SCALE}`;
+	do
+		make coverage-worker;
+	done
+fi
